Misc updates
- Se simplificó la tabla de Altas y Bajas de Personas
- Se renobraron y optimizaron las funciones de Sql y los stored procedures

+++ Desktop Application/usp_Personas_Result.vb
@@ -19,6 +19,6 @@ Partial Public Class usp_Personas_Result
     Public Property IDCuartel As Byte
     Public Property CuartelNombre As String
     Public Property EstadoActual As String
-    Public Property IDBajaMotivo As Nullable(Of Integer)
+    Public Property IDBajaMotivo As Nullable(Of Byte)
 
 End Class
